+++ src/lib.rs
@@ -1,6 +1,11 @@
 // Copyright (C) 2021 Daniel Mueller <deso@posteo.net>
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 
+#[cfg(not(target_arch = "wasm32"))]
+pub use hyper;
+#[cfg(target_arch = "wasm32")]
+pub use web_sys;
+
 mod error;
 mod issue;
 #[cfg(not(target_arch = "wasm32"))]