+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 ----------
 - Introduced `Issue` trait for issuing requests
 - Deprecated `Client` type
+- Added re-export for `hyper` and `web-sys` dependencies
 - Fixed compilation error when using native client